Add calibration-parser delegation in reference AST builder
This adds a search for through the entry point
`openqasm3.parser.calibration` for handlers for different calibration
grammars.  This allows other packages to register themselves as plugins,
and supply parsers for arbitrary `defcalgrammar` grammars.  The
`visitCalStatement` and `visitDefcalStatement` rules are updated to call
out to a suitable plugin class, which is loaded when a `defcalgrammar`
statement is encountered.

The intent is to produce the most general reference OpenQASM 3 parser,
and to be useful as a base for quick prototype tooling, or places where
the greatest speed is not as critical.  Highly performant parsers will
likely want to parse OpenQASM 3 and their chosen calibration grammar
with a single parser.

The current interface of the `CalibrationParser` abstract base class
does not include an item for passing a symbol table from the controlling
parser down to the subparsers.  In a more complete AST generator, this
would likely be required, but right now, the package does not do scope-
or type-system validation, so does not have such a symbol-table object
to pass.  We may well want to add this in the future.